+++ packages/polaris-viz-core/src/index.ts
@@ -106,6 +106,7 @@ export {
   getClosestDivisibleNumber,
   roundToDecimals,
   isLargeDataSet,
+  ColorScale,
 } from './utilities';
 export {
   useSparkBar,